requirements.txt
discord.py>=2.3
kubernetes>=29.0
boto3>=1.34
python-dotenv>=1.0
fastapi>=0.115
uvicorn>=0.30
aiohttp>=3.9
